refactor: add `monitor` module and move `InterestRegistry` into the core library
