+++ crates/argus-core/src/lib.rs
@@ -7,5 +7,6 @@ pub mod action_dispatcher;
 pub mod config;
 pub mod loader;
 pub mod models;
+pub mod monitor;
 pub mod persistence;
 pub mod providers;
