+++ src/monitor/mod.rs
@@ -5,7 +5,7 @@ mod manager;
 mod persistence_validator;
 mod validator;
 
-pub use interest_registry::{InterestRegistry, InterestRegistryBuilder};
+pub use interest_registry::InterestRegistryBuilder;
 pub use manager::{ClassifiedMonitor, MonitorAssetState, MonitorCapabilities, MonitorManager};
 pub use persistence_validator::{MonitorPersistenceValidationError, MonitorPersistenceValidator};
 pub use validator::{MonitorValidationError, MonitorValidator};
